+++ js/front/post.js
@@ -7,13 +7,16 @@ import log from 'logger';
 
 import Vue from 'vue';
 
+import 'less/post.less';
+
 // Components
 import ShareButton from 'components/buttons/share.vue';
+import DiscussionThreads from 'components/discussions/threads.vue';
 
 
 new Vue({
     mixins: [FrontMixin],
-    components: {ShareButton},
+    components: {ShareButton, DiscussionThreads},
     ready() {
         log.debug('Post page ready');
     }